requirement.txt
pywin32
pycrypto
requests